Atualizado as entidades do projeto

+++ src/main/java/br/senac/requestfood/model/consumivel/prato/Prato.java
@@ -20,7 +20,7 @@ public class Prato extends Consumivel {
 
     public Prato() {}
 
-    public Prato(Long id, Estabelecimento estabelecimento, String nome, float valor, String descricao, Byte[] imagem, CategoriaPrato tipoPrato){
+    public Prato(Long id, Estabelecimento estabelecimento, String nome, Double valor, String descricao, Byte[] imagem, CategoriaPrato tipoPrato){
         super(id, estabelecimento, nome, valor, descricao, imagem);
         setTipoPrato(tipoPrato);
     }
